Use uint for message counts.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@3407 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ src/common.h
@@ -141,9 +141,9 @@ typedef enum wl_e { WL_M_UNKNOWN ='U',
 /* Represents the secondary data for a word key */
 typedef struct {
     int freq;
-    long good;
-    long bad;
-    double prob;
+    uint	good;
+    uint	bad;
+    double 	prob;
 } wordprop_t;
 
 extern void bf_abort(void);
